DacFx bacpac export (sqlpackage / SSMS wizard) runs end-to-end against the TDS endpoint: sysconfigures / SESSIONPROPERTY  / xp_msver-by-RPC with @optname filtering, OPTION (USE HINT), name-resolution errors now batch-abort cleanly instead of cascading mid-parse, ~45 catalog views/columns incl. TYPE_TABLE sys.objects rows, first-class sql_variant with the SSVARIANTTYPE wire form, canonical per-table index-id allocation, computed-column definition capture into sys.computed_columns.definition, EXISTS((SELECT…)), DBCC SHOW_STATISTICS WITH HISTOGRAM as a MIN-anchored multi-step histogram, DATALENGTH returns bigint for (MAX) operands, string-concat MAX-sentinel arithmetic no longer kills sessions, binary-family promotion (varbinary(N) vs (MAX)), and byte-identical geography/geometry CLR serialization over the UDT 0xF0 wire form.

+++ SqlServerSimulator.Tests/StringScalarFunctionTests.cs
@@ -141,11 +141,11 @@ public sealed class StringScalarFunctionTests
     /// </summary>
     [TestMethod]
     public void Replicate_VarcharMax_NoTruncation()
-        => AreEqual(10000, ExecuteScalar<int>("select datalength(REPLICATE(cast('a' as varchar(max)), 10000))"));
+        => AreEqual(10000L, ExecuteScalar<long>("select datalength(REPLICATE(cast('a' as varchar(max)), 10000))"));
 
     [TestMethod]
     public void Replicate_NvarcharMax_NoTruncation_DoubleBytes()
-        => AreEqual(20000, ExecuteScalar<int>("select datalength(REPLICATE(cast('a' as nvarchar(max)), 10000))"));
+        => AreEqual(20000L, ExecuteScalar<long>("select datalength(REPLICATE(cast('a' as nvarchar(max)), 10000))"));
 
     /// <summary>
     /// Probe-confirmed 2026-07-10 against a live FROM-source column: a
@@ -154,7 +154,7 @@ public sealed class StringScalarFunctionTests
     /// </summary>
     [TestMethod]
     public void Replicate_VarcharMaxColumn_NoTruncation()
-        => AreEqual(20000, ExecuteScalar<int>("""
+        => AreEqual(20000L, ExecuteScalar<long>("""
             create table t (v varchar(max));
             insert t values (replicate('a', 100));
             select datalength(REPLICATE(v, 200)) from t
@@ -166,7 +166,7 @@ public sealed class StringScalarFunctionTests
     /// </summary>
     [TestMethod]
     public void Replicate_NvarcharMaxColumn_NoTruncation_DoubleBytes()
-        => AreEqual(40000, ExecuteScalar<int>("""
+        => AreEqual(40000L, ExecuteScalar<long>("""
             create table t (v nvarchar(max));
             insert t values (replicate(N'a', 100));
             select datalength(REPLICATE(v, 200)) from t
